Fix crash during deallocation of self-referencing lists
When a list or any of its nodes references itself in a way that creates a cycle,
their PyObjects may be visited multiple times during deallocation of the list.
To handle this case cleanly, Python documentation recommends to release
references held by such PyObject with Py_CLEAR.
The macro works similarly to Py_XDECREF, but also sets the reference to NULL,
which prevents repeated deallocation of the reference during subsequent visits
of its owner PyObject.

Most of the references held by llist PyObjects are released by Py_CLEAR.
One of the exceptions was Py_None, which is a global variable that may not be
set to NULL.
To work around this limitation, llist module simply invoked Py_DECREF on Py_None
during each visit of a garbage collected PyObject. Unfortunately this could
release Py_None too many times, effectively 'stealing' references from other
objects and eventually leading to a crash.

To fix this issue, each linked list and node which holds a Py_None reference
will now set a flag in its PyObject structure. The first release of the Py_None
reference will clear the flag to prevent future traversal of this object from
repeating the operation.

Closes issue #11.