Fix missing items during mutating list iteration
Iteration of sllist and dllist would miss an item if it had been added as the
successor of the current item in the previous step of the loop.

The issue was caused by the fact that the reference to next node was resolved in
the preceding step of the iteration.

With the fix introduced in this commit, a list node becomes current in the
iteration step when its value is returned.

Closes issue #12.