Fix tests checking recount on immortal objects
Starting with Python 3.12 singleton objects like `None` are considered
to be immortal. Their reference count is always set to a high value
(currently -1U).

In effect, some of the unit tests which verify ref counting of the None
object attempted to execute loops with ~43B iterations.

The test cases have been rewritten to validate the reference count with
a constant number of iterations.
The new implementation requires that `gc.collect` releases all reachable
references. This appears to be the case in all tested CPython versions.

Closes #18.